% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layers.R
\name{add_headers}
\alias{add_headers}
\title{Add headers to groups on a mosaic plot}
\usage{
add_headers(df, column = "f", cex, ...)
}
\arguments{
\item{df}{\code{mosaic_df}}

\item{column}{(\code{f} by default) name or position of a column in \code{df} to feed \code{\link[=lines]{lines()}}.
If a non-character column is provided, its \code{names}.}

\item{cex}{\code{numeric} to feed \code{\link[=text]{text()}}, if missing tries
to obtain something adequate}

\item{...}{additional arguments to feed \code{\link[=text]{text()}}}
}
\value{
the \code{mosaic_df}, invisibly
}
\description{
Add headers to groups on a mosaic plot
}
