% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layers.R
\name{add_lines}
\alias{add_lines}
\alias{add_curves}
\title{Add curves/lines to a mosaic plot}
\usage{
add_lines(df, column = "shp", col = rep(par("fg"), nrow(df)), lwd = rep(1,
  nrow(df)), lty = rep(par("lty"), nrow(df)), ...)

add_curves(df, column = "shp", col = rep(par("fg"), nrow(df)),
  lwd = rep(1, nrow(df)), lty = rep(par("lty"), nrow(df)), ...)
}
\arguments{
\item{df}{\code{mosaic_df}}

\item{column}{(\code{shp} by default) name or position of a
column in \code{df} to feed \code{\link[=lines]{lines()}}.}

\item{col}{\code{color} (recycled) to feed \code{\link[=lines]{lines()}}}

\item{lwd}{\code{numeric} (recycled) to feed \code{\link[=lines]{lines()}}}

\item{lty}{\code{numeric} (recycled) to feed \code{\link[=lines]{lines()}}}

\item{...}{additional arguments to feed \code{\link[=lines]{lines()}}}
}
\value{
the \code{mosaic_df}, invisibly
}
\description{
Add curves/lines to a mosaic plot
}
