% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layers.R
\name{add_polygons}
\alias{add_polygons}
\alias{add_outlines}
\title{Add outlines/polygons to a mosaic plot}
\usage{
add_polygons(df, column = "shp", density = rep(NA, nrow(df)), angle = 45,
  border = rep(NA, nrow(df)), col = rep(par("fg"), nrow(df)),
  lty = rep(par("lty"), nrow(df)), ...)

add_outlines(df, column = "shp", density = rep(NA, nrow(df)), angle = 45,
  border = rep(NA, nrow(df)), col = rep(par("fg"), nrow(df)),
  lty = rep(par("lty"), nrow(df)), ...)
}
\arguments{
\item{df}{\code{mosaic_df}}

\item{column}{(\code{shp} by default) name or position of a
column in \code{df} to feed \code{\link[=polygon]{polygon()}}.}

\item{density}{\code{numeric} (recycled) to feed \code{\link[=polygon]{polygon()}}}

\item{angle}{\code{numeric} (recycled) to feed \code{\link[=polygon]{polygon()}}}

\item{border}{\code{color} (recycled) to feed \code{\link[=polygon]{polygon()}}}

\item{col}{\code{color} (recycled) to feed \code{\link[=polygon]{polygon()}}}

\item{lty}{\code{numeric} (recycled) to feed \code{\link[=polygon]{polygon()}}}

\item{...}{additional arguments to feed \code{\link[=polygon]{polygon()}}}
}
\description{
Add outlines/polygons to a mosaic plot
}
