% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layers.R
\name{add_text}
\alias{add_text}
\alias{add_labels}
\title{Add text to shapes of a mosaic plot}
\usage{
add_text(df, column = "shp", col = rep(par("fg"), nrow(df)), font = rep(1,
  nrow(df)), cex, ...)

add_labels(df, column = "shp", col = rep(par("fg"), nrow(df)),
  font = rep(1, nrow(df)), cex, ...)
}
\arguments{
\item{df}{\code{mosaic_df}}

\item{column}{(\code{shp} by default) name or position of a
column in \code{df} to feed \code{\link[=lines]{lines()}}.
If a non-character column is provided, its \code{names}.}

\item{col}{\code{color} (recycled) to feed \code{\link[=text]{text()}}}

\item{font}{\code{numeric} (recycled) to feed \code{\link[=text]{text()}}}

\item{cex}{\code{numeric} to feed \code{\link[=text]{text()}}, if missing tries
to obtain something adequate}

\item{...}{additional arguments to feed \code{\link[=text]{text()}}}
}
\value{
the \code{mosaic_df}, invisibly
}
\description{
Add text to shapes of a mosaic plot
}
