% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momosaic-package.R
\docType{package}
\name{Momocs}
\alias{Momocs}
\alias{Momocs-package}
\title{Momosaics}
\description{
Use primitive layers to draw many shapes or
subplot as mosaics of plots.
}
\seealso{
\itemize{
\item \bold{Homepage}: \url{https://github.com/vbonhomme/Momosaics}
\item \bold{Issues}: \url{https://github.com/vbonhomme/Momosaics/issues}
\item \bold{Tutorial}: \code{browseVignettes("Momosaics")}
\item \bold{Email}: \code{bonhomme.vincent@gmail.com}
}
}
