% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mosaic_empty_plot}
\alias{mosaic_empty_plot}
\title{Given a mosaic_df, prepares the empty graphical window}
\usage{
mosaic_empty_plot(df)
}
\arguments{
\item{df}{a \code{mosaic_df}}
}
\description{
Given a mosaic_df, prepares the empty graphical window
}
