% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{mosaic_wrap}
\alias{mosaic_wrap}
\alias{mosaic_wrap_factor}
\title{Wrap a data.frame into a mosaic_df, with options}
\usage{
mosaic_wrap(df, nr, nc)

mosaic_wrap_factor(df, f, nr, nc)
}
\arguments{
\item{df}{\code{data.frame} with \code{shp} column as a \code{list} of shapes, and
possibly other columns (eg a \code{factor}) or directly a \code{list} of shapes}

\item{nr}{\code{integer} the number of rows for the mosaic plot}

\item{nc}{\code{interger} the number of columns for the mosaic plot}

\item{f}{\code{factor} for grouping structure}
}
\description{
Given a list of shapes or directly an existing \code{data.frame},
calculates loci positions (\code{row} and \code{col} columns, for their ids)
then decuces real coordinates (\code{xc} and \code{yc} columns).
}
\details{
If no \code{nr} (and/or \code{nc}) is passed, tries to return a squared mosaic plot.
}
\examples{
1:10 \%>\% mosaic_wrap()
1:10 \%>\% mosaic_wrap(nr=2)
}
