% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domestic.R
\name{palettize}
\alias{palettize}
\title{Creates a color vector from a factor and a palette name}
\usage{
palettize(f, palette)
}
\arguments{
\item{f}{\code{factor} (or another type of vector coerced as \code{factor})
as a grouping informations for colors to be returned}

\item{palette}{a color palette that, given a \code{integer} return a
color code.}
}
\description{
Can be used in conjunction with \code{add_*} functions.
}
\examples{
# creates a palette
w2b <- colorRampPalette(c("#FFFFFF", "#000000"))
# displays it
barplot(rep(1, 5), col=w2b(5))
# with palettize
palettize(iris$Species, w2b)
# alternatively you can use existing palettes
palettize(iris$Species, Momocs::col_summer)
}
