% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domestic.R
\name{transplate}
\alias{transplate}
\title{Template and translate shapes from a mosaic_df}
\usage{
transplate(df, size = 0.95)
}
\arguments{
\item{df}{a \code{mosaic_df} data.frame}

\item{size}{the templating size. See \code{\link[Momocs:coo_template]{Momocs::coo_template()}}}
}
\description{
Passed with a \code{mosaic_df}, template shapes to the desired size,
then translate shapes to the required locus using \code{xc} and \code{yc}
columns from the \code{data.frame}. Used internally by \code{wrap_*} functions.
}
